chore: port shared/AbstractEvent to TypeScript
