--- src/shared/AbstractEvent/index.js
@@ -1,3 +1,0 @@
-import AbstractEvent from './AbstractEvent';
-
-export default AbstractEvent;
